% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.R
\name{bw}
\alias{bw}
\title{Calculate the bw index}
\usage{
bw(distance, clusters, bwBalance = 1, indCluster = FALSE)
}
\arguments{
\item{distance}{Distance or Dis-Similarity Matrix}

\item{clusters}{The clustering to evaluate.}

\item{bwBalance}{The balance [0, 1] between inner cluster similarity (Compactness) and the similarity between clusters (Separation).
A balance value < 1 increases the importance of Compactness, whereas a value > 1 increases the importance of Separation.}

\item{indCluster}{If true, the bw value for all individual clusters is returned}
}
\value{
A score that describes how well the clustering fits the data.
}
\description{
Calculate the bw index
}
