% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterFunctions.R
\name{convert_columnwise_percent}
\alias{convert_columnwise_percent}
\title{Converts a matrix to relative abundances}
\usage{
convert_columnwise_percent(dat)
}
\arguments{
\item{dat}{The matrix to convert}
}
\value{
The matrix with all columns in percent
}
\description{
Converts a matrix to relative abundances
}
