% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{convert_columnwise_relative}
\alias{convert_columnwise_relative}
\title{Converts a matrix to relative abundances}
\usage{
convert_columnwise_relative(data)
}
\arguments{
\item{data}{A matrix of readouts that should be converted to relative
abundances}
}
\value{
The matrix with all columns in percent
}
\description{
Converts a matrix to relative abundances
}
