% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_clustering.R
\name{evaluate_clustering}
\alias{evaluate_clustering}
\title{Evaluate a clustering using the given method}
\usage{
evaluate_clustering(readouts, clustering, sim, method, custom_eval = NULL, ...)
}
\arguments{
\item{readouts}{The readouts the clustering and similarity matrix are based
on.}

\item{clustering}{The clustering to evaluate.}

\item{sim}{The similarity matrix, this clustering is based on.}

\item{method}{The method to evaluate the given clustering. This might be one
of "silhouette", "sdindex", "ptbiserial", "dunn", "bw", or
"custom'.}

\item{custom_eval}{A custom function to be run for evaluating a clustering.
Only used with method "custom".}

\item{...}{Further arguments that are passed to a custom function.}
}
\value{
A score that describes how well the clustering fits the data.
}
\description{
Evaluate a clustering using the given method
}
