% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_clustering.R
\name{evaluate_clustering_bw}
\alias{evaluate_clustering_bw}
\title{Evaluate a clustering using the bw index}
\usage{
evaluate_clustering_bw(readouts, clustering, sim, ...)
}
\arguments{
\item{readouts}{The readouts the clustering and similarity matrix are based
on.}

\item{clustering}{The clustering to evaluate.}

\item{sim}{The similarity matrix, this clustering is based on.}

\item{...}{Further arguments that are passed to the bw function.}
}
\value{
A score that describes how well the clustering fits the data.
}
\description{
Evaluate a clustering using the bw index
}
