% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_clustering.R
\name{evaluate_clustering_custom}
\alias{evaluate_clustering_custom}
\title{Evaluate a clustering using a custom evaluation function}
\usage{
evaluate_clustering_custom(readouts, clustering, sim, custom_eval, ...)
}
\arguments{
\item{readouts}{The readouts the clustering and similarity matrix are based
on.}

\item{clustering}{The clustering to evaluate.}

\item{sim}{The similarity matrix, this clustering is based on.}

\item{custom_eval}{The custom function to be run for evaluating a clustering.}

\item{...}{Further arguments that are passed to the custom function.}
}
\value{
A score that describes how well the clustering fits the data.
}
\description{
Evaluate a clustering using a custom evaluation function
}
