% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_clustering.R
\name{evaluate_clustering_dunn}
\alias{evaluate_clustering_dunn}
\title{Evaluate a clustering using the dunn index}
\usage{
evaluate_clustering_dunn(readouts, clustering, sim)
}
\arguments{
\item{readouts}{The readouts the clustering and similarity matrix are based
on.}

\item{clustering}{The clustering to evaluate.}

\item{sim}{The similarity matrix, this clustering is based on.}
}
\value{
A score that describes how well the clustering fits the data.
}
\description{
Evaluate a clustering using the dunn index
}
