% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateClustering.R
\name{evaluateClusteringBw}
\alias{evaluateClusteringBw}
\title{Evaluate a clustering using the bw index}
\usage{
evaluateClusteringBw(readouts, clustering, sim, ...)
}
\arguments{
\item{readouts}{The readouts the clustering and similarity matrix are based on.}

\item{clustering}{The clustering to evaluate.}

\item{sim}{The similarity matrix, this clustering is based on.}

\item{...}{Further arguments that are passed to the bw function.}
}
\value{
A score that describes how well the clustering fits the data.
}
\description{
Evaluate a clustering using the bw index
}
