% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateClustering.R
\name{evaluateClusteringSilhouette}
\alias{evaluateClusteringSilhouette}
\title{Evaluate a clustering using the silhouette index}
\usage{
evaluateClusteringSilhouette(readouts, clustering, sim)
}
\arguments{
\item{readouts}{The readouts the clustering and similarity matrix are based on.}

\item{clustering}{The clustering to evaluate.}

\item{sim}{The similarity matrix, this clustering is based on.}
}
\value{
A score that describes how well the clustering fits the data.
}
\description{
Evaluate a clustering using the silhouette index
}
