% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_at_tp_biggest_n}
\alias{filter_at_tp_biggest_n}
\title{Filters a matrix of readouts for the n biggest IS at a certain
measurement}
\usage{
filter_at_tp_biggest_n(data, at = "168", n = 50)
}
\arguments{
\item{data}{The readout matrix to filter.}

\item{at}{A filter for the columns/measurement. Only matching
columns/measurements are considered, though all will be returned.}

\item{n}{The number of biggest IS to return. If `at` matches multiple
columns/measurements, the 'rowSum()' over the columns/measurements
will be used. For ties, more than `n` IS may be returned.}
}
\value{
A matrix with only the n biggest IS at the selected measurements.
}
\description{
Filters a matrix of readouts for the n biggest IS at a certain
measurement
}
