% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterFunctions.R
\name{filter_atTP_biggestN}
\alias{filter_atTP_biggestN}
\title{Filters for the n biggest ISs}
\usage{
filter_atTP_biggestN(dat, at = "168", n = 50)
}
\arguments{
\item{dat}{The readout matrix to filter.}

\item{at}{A filter for the columns. Only (partially) matching columns are considered, though all will be returned.}

\item{n}{The number of biggest ISs to return. If `at` matches multiple columns, the rowSum will be used to rank the ISs. For ties, more than `n` ISs can be returned.}
}
\value{
A matrix with only the n biggest ISs at the selected measurements.
}
\description{
Filters for the n biggest ISs
}
