% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterFunctions.R
\name{filter_atTP_min}
\alias{filter_atTP_min}
\title{Filters for ISs that have a minimum occurence}
\usage{
filter_atTP_min(dat, at = "168", min = 0.02)
}
\arguments{
\item{dat}{The readout matrix to filter.}

\item{at}{A filter for the columns. Only (partially) matching columns are considered, though all will be returned.}

\item{min}{The minumum with which an IS has to occurr. This could be either absolute or relative reads. If `at` matches multiple columns, the rowSum will be used.}
}
\value{
A matrix with only the ISs that occurr with a minimum at the selected measurements.
}
\description{
Filters for ISs that have a minimum occurence
}
