% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_is_names}
\alias{filter_is_names}
\title{Shortens the rownames of a readout matrix to the shortest distinct prefix}
\usage{
filter_is_names(dat, by = "[_():]|[^_():]*")
}
\arguments{
\item{dat}{The readout matrix for which the names should be filtered.}

\item{by}{The regexp used to split the names.}
}
\value{
A matrix with the names filtered to the shortest unique prefix.
}
\description{
Shortens the rownames of a readout matrix to the shortest distinct prefix
}
\seealso{
filter.names
}
