% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_match}
\alias{filter_match}
\title{Filters for columns containing a certain substring.}
\usage{
filter_match(dat, match = "E2P11")
}
\arguments{
\item{dat}{The readout matrix to filter.}

\item{match}{The substring that columns must match.}
}
\value{
A readout matrix that only contains the columns whose names contain
        the substring.
}
\description{
Filters for columns containing a certain substring.
}
