% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_measurement_names}
\alias{filter_measurement_names}
\title{Splits a vector of strings by a given regexp, selects and rearranges the
parts and joins them again}
\usage{
filter_measurement_names(dat, elems = c(1, 3), by = "_")
}
\arguments{
\item{dat}{The readout matrix to filter.}

\item{elems}{The elements to select. They are rearrange in the order that is
given via this argument.}

\item{by}{The string used for splitting the names of the columns.}
}
\value{
A matrix where the names of the columns are split by the given
        string, rearranged and again joined by the string.
}
\description{
Splits a vector of strings by a given regexp, selects and rearranges the
parts and joins them again
}
