% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_names}
\alias{filter_names}
\title{Filters a vector of names and returns the shortest common prefix.}
\usage{
filter_names(names, by = "[_():]|[^_():]*")
}
\arguments{
\item{names}{The vector of names to filter.}

\item{by}{A regexp that splits the string. The default filters by special
characters. A split by character can be achieved by using "." as
the regexp.}
}
\value{
The names shortened to the shortest prefix (in chunks defined by
        the regexp) where all names are unique.
}
\description{
Filters a vector of names and returns the shortest common prefix.
}
