% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterFunctions.R
\name{filter_nrTP_min}
\alias{filter_nrTP_min}
\title{Filters for a minimum number of time points/measurements}
\usage{
filter_nrTP_min(dat, min = 6)
}
\arguments{
\item{dat}{The readout matrix to filter.}

\item{min}{The minimum number of measurements where an IS needs to have a value that is not 0 or NA.}
}
\value{
A matrix with only ISs that have more than `min` columns that are not 0 or NA.
}
\description{
Filters for a minimum number of time points/measurements
}
