% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_zero_columns}
\alias{filter_zero_columns}
\title{Removes columns that only contain 0 or NA.}
\usage{
filter_zero_columns(dat)
}
\arguments{
\item{dat}{The readout matrix to filter.}
}
\value{
A matrix where columns that where only 0 or NA are filtered out.
}
\description{
Removes columns that only contain 0 or NA.
}
