% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filter_zero_rows}
\alias{filter_zero_rows}
\title{Removes rows that only contain 0 or NA.}
\usage{
filter_zero_rows(dat)
}
\arguments{
\item{dat}{The readout matrix to filter.}
}
\value{
A matrix where rows that where only 0 or NA are filtered out.
}
\description{
Removes rows that only contain 0 or NA.
}
