% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBestNrCluster.R
\name{findBestNrCluster}
\alias{findBestNrCluster}
\title{Finds the best number of clusters according to silhouette}
\usage{
findBestNrCluster(
  data,
  sim,
  method.reconstruction = "kmedoids",
  method.evaluation = "silhouette",
  report = FALSE,
  parallel = FALSE,
  best = max,
  returnAll = FALSE,
  ...
)
}
\arguments{
\item{data}{The barcode data in a matrix.}

\item{sim}{A similarity matrix.}

\item{method.reconstruction}{The clustering method to use.}

\item{method.evaluation}{The evaluation method to use.}

\item{report}{Whether the current progress should be reported. Note that this will not work if parallel is set to TRUE.}

\item{parallel}{Whether the clustering should be performed in parallel.}

\item{best}{The method to use to determine the best clustering.}

\item{returnAll}{Wheter to return the silhouette score for all clusterings.}

\item{...}{passed params to evaluating clustering}
}
\value{
The R^2 value for rows is1 and is2 in matrix dat
}
\description{
Finds the best number of clusters according to silhouette
}
