% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.clusterObj}
\alias{plot.clusterObj}
\title{Plots the clustering based on a clustering object}
\usage{
\method{plot}{clusterObj}(x, ...)
}
\arguments{
\item{x}{The clustering object.}

\item{...}{Further arguments are ignored.}
}
\value{
A ggplot object, which can be used to further individualize or to
        plot directly.
}
\description{
Plots the clustering based on a clustering object
}
