% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.ISSimilarity}
\alias{plot.ISSimilarity}
\title{Plots the similarity of integration sites}
\usage{
\method{plot}{ISSimilarity}(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{The matrix that holds the similarity values}

\item{na.rm}{whether NA values should be deleted beforehand}

\item{...}{Further arguments are ignored.}
}
\value{
A ggplot object, which can be used to further individualize or to
        plot directly.
}
\description{
Plots the similarity of integration sites
}
