% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.timeseries}
\alias{plot.timeseries}
\title{Plots time series data, which consists of multiple measurements over
time / place (cols) of different clones / integration sites (rows).}
\usage{
\method{plot}{timeseries}(x, ...)
}
\arguments{
\item{x}{The data to plot.}

\item{...}{Further arguments are ignored.}
}
\value{
A ggplot object, which can be used to further individualize or to
        plot directly.
}
\description{
Plots time series data, which consists of multiple measurements over
time / place (cols) of different clones / integration sites (rows).
}
