% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_rsquare}
\alias{plot_rsquare}
\title{Plots R^2 of two integration sites}
\usage{
plot_rsquare(dat, is1, is2)
}
\arguments{
\item{dat}{The matrix that holds the values}

\item{is1}{The name of the first row}

\item{is2}{The name of the second row}
}
\value{
A ggplot object, which can be used to further individualize or to
        plot directly.
}
\description{
Plots R^2 of two integration sites
}
