% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{reconstruct_kmedoid}
\alias{reconstruct_kmedoid}
\title{Calculate the k-medoids clustering for a given time course.}
\usage{
reconstruct_kmedoid(
  readouts,
  target_communities,
  sim = MultIS::get_similarity_matrix(readouts = readouts, self = 0, upper = TRUE)
)
}
\arguments{
\item{readouts}{The time course for which to find clusters.}

\item{target_communities}{The number of clusters to cluster for.}

\item{sim}{A similarity matrix for the time course.}
}
\value{
A matrix with two columns: "Clone" and "IS".
}
\description{
Calculate the k-medoids clustering for a given time course.
}
