% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{reconstructKMedoid}
\alias{reconstructKMedoid}
\title{Calculate the k-medoids clustering for a given time course.}
\usage{
reconstructKMedoid(
  readouts,
  targetCommunities,
  sim = MultIS::getSimilarityMatrix(readouts = readouts, self = 0, upper = TRUE)
)
}
\arguments{
\item{readouts}{The time course for which to find clusters.}

\item{targetCommunities}{The number of clusters to cluster for.}

\item{sim}{A similarity matrix for the time course.}
}
\value{
A matrix with two columns: "Clone" and "Barcode".
}
\description{
Calculate the k-medoids clustering for a given time course.
}
